refactor: create execution/process/ subpackage
Move process.py → process/__init__.py and 6 _process_*.py modules into
execution/process/. Update all internal cross-references and the
execution/__init__.py direct import of _process_kill.
