+++ src/autoskillit/execution/__init__.py
@@ -6,7 +6,7 @@ execution/headless.py has TYPE_CHECKING-only references to pipeline/.
 """
 
 from autoskillit.core import SkillResult
-from autoskillit.execution._process_kill import (
+from autoskillit.execution.process._process_kill import (
     async_kill_process_tree,
     kill_process_tree,
 )
